Downgrade Scala compiler dependencies (PRESS-20110)
I believe we're currently upgrading our stack in a different order than
Lucid Software. We are upgrading to Java 11 before upgrading our Scala
version. We are still on Scala 2.13.3.

The version of scala_rules that we're branching from in this commit,
has fixes we need to run the compiler and tests with Java 11. At the
same time the version also uses compiler dependencies for Scala 2.13.6.
To try and make sure we just upgrade one thing at a time, I'm
downgrading compiler dependencies to be compatible with Scala
2.13.3

I was able to determine which version to use by checking Maven. Maven
has a section that lists "Compile Dependencies" for the compiler
dependencies [1]. I looked for the scala-library version that was
listed.

EDIT: It turned out the version I picked (1.4.4) had a bug. I backported
the bug fix. That also meant I had to compile the version from source
and publish it to our JFrog repository.

[1]: https://mvnrepository.com/artifact/org.scala-sbt/zinc_2.13/1.4.4